splitup record decoder
